test/Authentication/AuthenticationTest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit5TestClass.java to edit this template
 */
package Authentication;

import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.AfterAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

/**
 *
 * @author ntand
 */
public class AuthenticationTest {
    
    public AuthenticationTest() {
    }
    
    @BeforeAll
    public static void setUpClass() {
    }
    
    @AfterAll
    public static void tearDownClass() {
    }
    
    @BeforeEach
    public void setUp() {
    }
    
    @AfterEach
    public void tearDown() {
    }

    /**
     * Test of isUsernameValid method, of class Authentication.
     */
    @Test
    public void testIsUsernameValid() {
        System.out.println("isUsernameValid");
        Authentication instance = new Authentication();
        boolean expResult = false;
        boolean result = instance.isUsernameValid();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of isPasswordComplex method, of class Authentication.
     */
    @Test
    public void testIsPasswordComplex() {
        System.out.println("isPasswordComplex");
        Authentication instance = new Authentication();
        boolean expResult = false;
        boolean result = instance.isPasswordComplex();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of registerUser method, of class Authentication.
     */
    @Test
    public void testRegisterUser() {
        System.out.println("registerUser");
        Authentication instance = new Authentication();
        String expResult = "";
        String result = instance.registerUser();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of loginUser method, of class Authentication.
     */
    @Test
    public void testLoginUser() {
        System.out.println("loginUser");
        Authentication instance = new Authentication();
        boolean expResult = false;
        boolean result = instance.loginUser();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of returnLoginStatus method, of class Authentication.
     */
    @Test
    public void testReturnLoginStatus() {
        System.out.println("returnLoginStatus");
        Authentication instance = new Authentication();
        String expResult = "";
        String result = instance.returnLoginStatus();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of main method, of class Authentication.
     */
    @Test
    public void testMain() {
        System.out.println("main");
        String[] args = null;
        Authentication.main(args);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }
    
}
